+++ client/src/components/Header.module.css
@@ -1,0 +1,5 @@
+.logo {
+    width: 60px;
+    height: 60px;
+  }
+
