+++ client/src/pages/Home.module.css
@@ -1,0 +1,39 @@
+.backgroundImage {
+    background-image: url('../assets/homesec1.jpg');
+    background-size: cover;
+    background-position: center;
+  }
+  
+  .overlay {
+    background-color: rgba(0, 0, 0, 0.5);
+  }
+  
+  .title {
+    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
+  }
+  
+  .subtitle {
+    text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.5);
+  }
+  
+  .aboutSection {
+    background-color: #f3f4f6;
+  }
+  
+  .missionVisionCard {
+    transition: transform 0.3s ease-in-out;
+  }
+  
+  .missionVisionCard:hover {
+    transform: translateY(-5px);
+  }
+  
+  @media (max-width: 640px) {
+    .title {
+      font-size: 2rem;
+    }
+    
+    .subtitle {
+      font-size: 1rem;
+    }
+  }
