--- src/components/Header.module.css
@@ -1,4 +1,0 @@
-.logo {
-  width: 60px;
-  height: 60px;
-}
